+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,19 +1,19 @@
 package roomescape.domain;
 
 public class Reservation {
-    private final int id;
+    private final long id;
     private final String name;
     private final String date;
     private final String time;
 
-    public Reservation(int id, String name, String date, String time) {
+    public Reservation(long id, String name, String date, String time) {
         this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
